backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Dependencies
langchain==0.0.350
langgraph==0.0.25
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0

# Audio Processing
openai-whisper==20231117
librosa==0.10.1
soundfile==1.0.2
pydub==0.25.1

# TTS
TTS==0.20.6
piper-tts==1.2.0
espeak-ng==1.0.3

# Image Processing
Pillow==10.0.0
opencv-python==4.8.1.78

# HTTP Client
aiohttp==3.9.0
httpx==0.25.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
